Resize canvas correctly
The window resizes differently on different devices. The innerWidth and
innerHeight properties of the window aren't sufficient for the task, and
when rendering in regular mode it's necessary to read a devicePixelRatio
property. Finally, the HMD specifies recommended dimensions at which to
render, so if the VREffect is active, we should use that instead.

- Sean T. McBeth <sean.mcbeth@gmail.com>